APPSRE-10229 DTP cleanup
