+++ reconcile/cli.py
@@ -2953,20 +2953,20 @@ def cluster_auth_rhidp(ctx):
 )
 @click.option(
     "--ocm-org-ids",
-    help="A comma seperated list of OCM organization IDs DTP should operator on. If none is specified, all organizations are considered.",
+    help="A comma seperated list of OCM organization IDs DTP should operate on. If none is specified, all organizations are considered.",
     required=False,
     envvar="DTP_OCM_ORG_IDS",
 )
 @click.pass_context
 def dynatrace_token_provider(ctx, ocm_org_ids):
-    from reconcile import dynatrace_token_provider
-    from reconcile.dynatrace_token_provider import (
+    from reconcile.dynatrace_token_provider.integration import (
+        DynatraceTokenProviderIntegration,
         DynatraceTokenProviderIntegrationParams,
     )
 
     parsed_ocm_org_ids = set(ocm_org_ids.split(",")) if ocm_org_ids else None
     run_class_integration(
-        integration=dynatrace_token_provider.DynatraceTokenProviderIntegration(
+        integration=DynatraceTokenProviderIntegration(
             DynatraceTokenProviderIntegrationParams(
                 ocm_organization_ids=parsed_ocm_org_ids
             )
