+++ reconcile/dynatrace_token_provider/merics.py
@@ -1,0 +1,43 @@
+from pydantic import BaseModel
+
+from reconcile.utils.metrics import (
+    CounterMetric,
+    ErrorRateMetricSet,
+)
+
+
+class DTPBaseMetric(BaseModel):
+    integration: str
+    ocm_env: str
+
+
+class DTPOrganizationReconcileCounter(DTPBaseMetric, CounterMetric):
+    org_id: str
+
+    @classmethod
+    def name(cls) -> str:
+        return "dtp_organization_reconciled"
+
+
+class DTPOrganizationReconcileErrorCounter(DTPBaseMetric, CounterMetric):
+    org_id: str
+
+    @classmethod
+    def name(cls) -> str:
+        return "dtp_organization_reconcile_errors"
+
+
+class DTPOrganizationErrorRate(ErrorRateMetricSet):
+    def __init__(self, integration: str, org_id: str, ocm_env: str) -> None:
+        super().__init__(
+            counter=DTPOrganizationReconcileCounter(
+                integration=integration,
+                ocm_env=ocm_env,
+                org_id=org_id,
+            ),
+            error_counter=DTPOrganizationReconcileErrorCounter(
+                integration=integration,
+                ocm_env=ocm_env,
+                org_id=org_id,
+            ),
+        )
